extract delegated capacity function to FimHelper
