+++ fabric_cf/actor/core/policy/network_service_inventory.py
@@ -29,6 +29,7 @@ import traceback
 from ipaddress import IPv6Network, IPv4Network
 from typing import List, Tuple, Union
 
+from fabric_cf.actor.fim.fim_helper import FimHelper
 from fim.slivers.capacities_labels import Labels
 from fim.slivers.gateway import Gateway
 from fim.slivers.interface_info import InterfaceSliver, InterfaceType
@@ -333,7 +334,7 @@ class NetworkServiceInventory(InventoryForType):
                 return requested_ns
 
             # Grab Label Delegations
-            delegation_id, delegated_label = self.get_delegations(delegations=owner_ns.get_label_delegations())
+            delegation_id, delegated_label = FimHelper.get_delegations(delegations=owner_ns.get_label_delegations())
 
             # HACK to use FabNetv6 for FabNetv6Ext as both have the same range
             requested_ns_type = requested_ns.get_type()
@@ -564,7 +565,7 @@ class NetworkServiceInventory(InventoryForType):
             ifs_labels = Labels()
 
         if owner_switch.get_name() == Constants.AL2S:
-            delegation_id, delegated_label = self.get_delegations(delegations=bqm_interface.get_label_delegations())
+            delegation_id, delegated_label = FimHelper.get_delegations(delegations=bqm_interface.get_label_delegations())
             local_name = delegated_label.local_name
             device_name = delegated_label.device_name
         else:
