+++ fabric_cf/orchestrator/core/resource_tracker.py
@@ -2,7 +2,8 @@ from collections import defaultdict
 from datetime import datetime, timedelta, timezone
 import logging
 
-from fabric_cf.actor.core.policy.network_node_inventory import NetworkNodeInventory
+from fabric_cf.actor.fim.fim_helper import FimHelper
+
 from fim.slivers.base_sliver import BaseSliver
 
 from fabric_cf.actor.core.plugins.db.actor_database import ActorDatabase
@@ -82,7 +83,7 @@ class ResourceTracker:
         :param cbm_node: The CBM node from which to initialize capacities and components.
         :type cbm_node: NodeSliver
         """
-        _, self.total_capacities = NetworkNodeInventory.get_delegations(delegations=cbm_node.get_capacity_delegations())
+        _, self.total_capacities = FimHelper.get_delegations(delegations=cbm_node.get_capacity_delegations())
         self.total_components = {}
 
         if cbm_node.attached_components_info:
